requirements.txt
selenium==4.12.0
pytest==7.4.2
allure-pytest==2.13.3
requests==2.31.0
webdriver-manager==3.8.6
pytest-xdist==3.3.1
pytest-html==4.1.1
